requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
plotly>=5.18.0
gspread>=6.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
pytest>=7.4.0
python-dateutil>=2.8.2